+++ Core/Entity/Review.cs
@@ -9,11 +9,11 @@ namespace MyNeighbors.Core.Entity
     {
         public int Id { get; set; }
         public string Description { get; set; } 
-        public double Rating { get; set; }
+        public int Rating { get; set; }
         public DateTime Date { get; set; }
-        public double Noise_Rating { get; set; }
-        public double Shopping_Rating { get; set; }
-        public double Schools_Rating { get; set; }
+        public int Noise_Rating { get; set; }
+        public int Shopping_Rating { get; set; }
+        public int Schools_Rating { get; set; }
         public int UserId { get; set; }
 
     }
